refactor: extract prismctl client package for better code organization
- Create cmd/prismctl/client/ package with API types and methods
- Move PrismAPIClient and all API response types to client package
- Update main.go to import and use client package types
- Remove duplicate API code from main.go (682 lines removed)
- Reduce main.go from 1703 to 1021 lines (40% reduction)
- Maintain all existing functionality with cleaner separation of concerns
